Started moving the etcd logic into etcd folder
