+++ config/config.json
@@ -1,5 +1,6 @@
 {
     "RequiresAuth": false,
+    "dataSource": "etcd",
     "etcdHost": "127.0.0.1",
     "etcdPort": "4001",
     "hobknobHost": "localhost",
